Init new screens w/window size + drag editor map
Screens are initialized with the current window size to avoid hard
coding default window sizes and then resizing (although resize is still
called when a new screen is started, no matter what).

SquareEditor has two cameras/viewports. Took me quite a while to figure
out how to properly isolate them, and place things in the right area.
(For the longest time the map in the background was part of the UI,
despite the printed text correctly being in the game.)

If you hold down the right mouse button and move the mouse, the game/map
will move with it. Next plan is to add a drop down menu on right click
(not dragging of course).